requirements.txt
openai>=1.0.0
pinecone>=5.0.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
tqdm>=4.64.0
firebase-admin>=6.0.0